-Transfer methods Details, Share from HomeController to MatchController -Add fild in MatchController, and fix this constructor -Fix Share method in MatchController -Add MatchSharedDetailsViewModel class -Fix Home/Index view -Add Match/Share view -Fix Match/Details view
