+++ src/YoTennis/Views/Home/Index.cshtml
@@ -136,7 +136,7 @@
                         </td>
                         <td>
                             @Html.ActionLink("Edit", "Index", "Match", new { id = match.Id }) |
-                            @Html.ActionLink("Details", "Details", new { id = match.Id }) |
+                            @Html.ActionLink("Details", "Details", "Match", new { id = match.Id }) |
                             @Html.ActionLink("Stats", "Stats", "Match", new { id = match.Id }) |
                             @Html.ActionLink("Delete", "Delete", "Home", new
                        {
